zeek-live/data-agent/requirements.txt
flask
requests
gunicorn
python-keycloak
